Refactor AI interaction logic, update UI layouts for compact states, and improve localization.
- **AI Integration**: Implement a local fallback mechanism in `OllamaAiRepository` for failed primary API requests.
- **Weather Story**: Add a manual "Generate" button for weather stories and implement numeric rounding for displayed values.
- **UI Improvements**: Add `compact` mode support to `SkyAnalyzerScreen` components to adjust layout when a photo is present; update branding from "Weather Simulator AI" to "Stormy" in Auth screens.
- **Localization**: Comprehensive update of hardcoded Romanian text to include proper diacritics across Games, Quiz, and Nature modules.
- **Logic Fixes**: Adjust `NatureImpactScreen` scoring to properly zero out winter-specific signals during warm conditions; update `WeatherStoryRequest` and `WeatherViewModel` to use `Int` instead of `Double` for cleaner AI prompts.
- **Bug Fixes**: Reset weather story state when new weather data is applied.

+++ app/src/main/java/com/example/weathersimulator/data/remote/ai/WeatherStoryRequest.kt
@@ -1,20 +1,20 @@
 package com.example.weathersimulator.data.remote.ai
 
 data class WeatherStoryRequest(
-    val temperature: Double?,
-    val apparentTemperature: Double?,
+    val temperature: Int?,
+    val apparentTemperature: Int?,
     val humidity: Int?,
-    val windSpeed: Double?,
-    val pressure: Double?,
+    val windSpeed: Int?,
+    val pressure: Int?,
     val weatherCode: Int?,
     val cloudCover: Int?,
     val nextHours: List<String>,
-    val nextTemperatures: List<Double>,
-    val nextPrecipitation: List<Double>,
-    val nextWindSpeed: List<Double>,
+    val nextTemperatures: List<Int>,
+    val nextPrecipitation: List<Int>,
+    val nextWindSpeed: List<Int>,
     val nextWeatherCodes: List<Int>,
     val dailyDates: List<String>,
-    val dailyMaxTemperatures: List<Double>,
-    val dailyMinTemperatures: List<Double>,
+    val dailyMaxTemperatures: List<Int>,
+    val dailyMinTemperatures: List<Int>,
     val dailyWeatherCodes: List<Int>
 )
